+++ repokid/datasource/__init__.py
@@ -11,57 +11,3 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
-
-import logging
-from typing import Dict
-from typing import Generic
-from typing import ItemsView
-from typing import Iterable
-from typing import Iterator
-from typing import KeysView
-from typing import Optional
-from typing import TypeVar
-from typing import ValuesView
-from typing import cast
-
-from repokid.plugin import RepokidPlugin
-from repokid.types import RepokidConfig
-
-logger = logging.getLogger("repokid")
-KT = TypeVar("KT")
-VT = TypeVar("VT")
-
-
-class DatasourcePlugin(RepokidPlugin, Generic[KT, VT]):
-    """A dict-like container that can be used to retrieve and store data"""
-
-    _data: Dict[KT, VT]
-
-    def __init__(self, config: Optional[RepokidConfig] = None):
-        super().__init__(config=config)
-        self._data = {}
-
-    def __getitem__(self, name: KT) -> VT:
-        return self._data[name]
-
-    def __iter__(self) -> Iterator[VT]:
-        return iter(cast(Iterable[VT], self._data))
-
-    def keys(self) -> KeysView[KT]:
-        return self._data.keys()
-
-    def items(self) -> ItemsView[KT, VT]:
-        return self._data.items()
-
-    def values(self) -> ValuesView[VT]:
-        return self._data.values()
-
-    def get(self, identifier: KT) -> VT:
-        raise NotImplementedError
-
-    def seed(self, identifier: KT) -> None:
-        raise NotImplementedError
-
-    def reset(self) -> None:
-        logger.debug("resetting %s", type(self).__name__)
-        self._data = {}
