+++ tests/datasource/conftest.py
@@ -2,10 +2,10 @@ from typing import List
 
 import pytest
 
-from repokid.datasource import DatasourcePlugin
 from repokid.datasource.access_advisor import AccessAdvisorDatasource
 from repokid.datasource.iam import ConfigDatasource
 from repokid.datasource.iam import IAMDatasource
+from repokid.datasource.plugin import DatasourcePlugin
 
 
 @pytest.fixture(autouse=True)
